requirements.txt
fastapi
uvicorn[standard]
httpx
google-genai
pydantic-settings
